+++ test/with-packages/default.nix
@@ -5,7 +5,7 @@ with util;
 
 let
   pkgs = import ./pkgs.nix;
-  pkgSet = mkPkgSet {
+  pkgSet = dec: mkPkgSet {
     # generated with:
     #   cabal new-build
     #   plan-to-nix -o .
@@ -19,65 +19,64 @@ let
 
       # vary component config for tests
       {
-        packages.test-with-packages.components = {
-          all.doExactConfig = mkForce false; # the default; use force because of quirk in all. See package.nix
-          library.doExactConfig = true; # not the default
-        };
+        packages.test-with-packages.components.library.doExactConfig = dec;
       }
     ];
   };
 
-  packages = pkgSet.config.hsPkgs;
+  packages = dec: (pkgSet dec).config.hsPkgs;
 
-  package = packages.test-with-packages;
-  inherit (package.components) library;
+  package = dec: (packages dec).test-with-packages;
+
+  decLibrary = (package true).components.library;
+  library = (package false).components.library;
 
   pkgId = p: "${p.identifier.name}-${p.identifier.version}";
-  showDepends = component: concatMapStringsSep " " pkgId component.depends;
+  showDepends = component: concatMapStringsSep " " pkgId (component.depends or []);
 
 in recurseIntoAttrs {
   # Used for testing externally with nix-shell (../tests.sh).
   # This just adds cabal-install to the existing shells.
-  test-shell = addCabalInstall packages.test-with-packages.components.all;
+  test-shell = addCabalInstall library;
 
   # A variant of test-shell with the component option doExactConfig enabled
-  test-shell-dec = addCabalInstall packages.test-with-packages.components.library;
+  test-shell-dec = addCabalInstall decLibrary;
 
   run = stdenv.mkDerivation {
     name = "with-packages-test";
-    libraryDepends = showDepends pkgSet.config.packages.test-with-packages.components.library;
-    allDepends = showDepends pkgSet.config.packages.test-with-packages.components.all;
+    decLibraryDepends = showDepends (pkgSet true).config.packages.test-with-packages.components.library;
+    libraryDepends = showDepends (pkgSet false).config.packages.test-with-packages.components.library;
 
     buildCommand = ''
       ########################################################################
       # test with-packages
 
-      printf "checking merging of the 'all' component depends ... " >& 2
-      if [ -n "$libraryDepends" -a "$libraryDepends" = "$allDepends" ]; then
+      printf "checking component depends ... " >& 2
+      if [ -n "$decLibraryDepends" -a "$decLibraryDepends" = "$libraryDepends" ]; then
         echo "PASS" >& 2
       else
         echo "FAIL" >& 2
+        echo "decLibraryDepends = $decLibraryDepends"
         echo "libraryDepends = $libraryDepends"
-        echo "allDepends = $allDepends"
         exit 1
       fi
 
-      printf "checking that the 'all' component works... " >& 2
-      echo ${package.components.all} >& 2
+      printf "checking that the 'library' without doExactConfig works... " >& 2
+      echo ${library} >& 2
 
-      printf "checking that the 'library' component works... " >& 2
-      echo ${package.components.library} >& 2
+      printf "checking that the 'library' with doExactConfig works... " >& 2
+      echo ${decLibrary} >& 2
     '' + (if stdenv.hostPlatform.isWindows
       then ''
         printf "runghc tests are not working yet for windows. skipping. " >& 2
       ''
       else ''
-        printf "checking that the package env has the dependencies... " >& 2
-        ${package.components.all.env}/bin/${package.components.all.env.targetPrefix}runghc ${./Point.hs}
+        printf "checking that non doExactConfig liarary.env has the dependencies... " >& 2
+        ${library.env}/bin/${library.env.targetPrefix}runghc ${./Point.hs}
         echo >& 2
 
-        printf "checking that components.library.env has the dependencies... " >& 2
-        ${library.env}/bin/${library.env.targetPrefix}runghc ${./Point.hs}
+        printf "checking that doExactConfig library.env has the dependencies... " >& 2
+        ${decLibrary.env}/bin/${decLibrary.env.targetPrefix}runghc ${./Point.hs}
         echo >& 2
       '') + ''
       touch $out